--- src/Lewee.Infrastructure.AspNet.WebApi/CommandType.cs
@@ -1,27 +1,0 @@
-﻿namespace Lewee.Infrastructure.AspNet.WebApi;
-
-/// <summary>
-/// Command Type
-/// </summary>
-public enum CommandType
-{
-    /// <summary>
-    /// Post
-    /// </summary>
-    Post,
-
-    /// <summary>
-    /// Put
-    /// </summary>
-    Put,
-
-    /// <summary>
-    /// Patch
-    /// </summary>
-    Patch,
-
-    /// <summary>
-    /// Delete
-    /// </summary>
-    Delete
-}
