--- src/Lewee.Infrastructure.AspNet.WebApi/Lewee.Infrastructure.AspNet.WebApi.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  
-  <ItemGroup>
-    <PackageReference Include="FastEndpoints" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\Lewee.Infrastructure.AspNet\Lewee.Infrastructure.AspNet.csproj" />
-  </ItemGroup>
-  
-</Project>
