+++ src/Lewee.Infrastructure.FastEndpoints/CommandType.cs
@@ -1,0 +1,27 @@
+﻿namespace Lewee.Infrastructure.FastEndpoints;
+
+/// <summary>
+/// Command Type
+/// </summary>
+public enum CommandType
+{
+    /// <summary>
+    /// Post
+    /// </summary>
+    Post,
+
+    /// <summary>
+    /// Put
+    /// </summary>
+    Put,
+
+    /// <summary>
+    /// Patch
+    /// </summary>
+    Patch,
+
+    /// <summary>
+    /// Delete
+    /// </summary>
+    Delete
+}
