+++ src/Lewee.Infrastructure.FastEndpoints/CorrelationContextAccessorExtensions.cs
@@ -1,0 +1,18 @@
+﻿using Correlate;
+
+namespace Lewee.Infrastructure.FastEndpoints;
+
+internal static class CorrelationContextAccessorExtensions
+{
+    public static Guid GetCorrelationId(this ICorrelationContextAccessor correlationContextAccessor)
+    {
+        var cid = correlationContextAccessor?.CorrelationContext?.CorrelationId;
+
+        if (cid == null || !Guid.TryParse(cid, out var correlationId))
+        {
+            return Guid.NewGuid();
+        }
+
+        return correlationId;
+    }
+}
